+++ pkg/client/logged/client.go
@@ -1,4 +1,4 @@
-package client
+package logged
 
 import (
 	"context"
