+++ pkg/client/logged/client_test.go
@@ -1,4 +1,4 @@
-package client
+package logged
 
 import (
 	"context"
