+++ pkg/runner/clusters/client.go
@@ -3,8 +3,8 @@ package clusters
 import (
 	"github.com/kyverno/chainsaw/pkg/client"
 	"github.com/kyverno/chainsaw/pkg/client/dryrun"
+	"github.com/kyverno/chainsaw/pkg/client/logged"
 	"github.com/kyverno/chainsaw/pkg/client/simple"
-	runnerclient "github.com/kyverno/chainsaw/pkg/runner/client"
 	"k8s.io/client-go/rest"
 )
 
@@ -13,7 +13,7 @@ func makeClient(config *rest.Config, dryRun bool) (client.Client, error) {
 	if err != nil {
 		return nil, err
 	}
-	c = runnerclient.New(c)
+	c = logged.New(c)
 	if !dryRun {
 		return c, nil
 	}
